pyproject.toml
[project]
name = "picsconvert"
version = "0.1.0"
description = "图片格式转换工具，支持批量处理压缩包中的图片"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.0",
    "pillow>=10.0.0",
    "pyvips>=2.2.1",
    "textual>=0.40.0",
    "pathlib>=1.0.1",
    "typing-extensions>=4.7.0",
    "pillow-avif-plugin>=1.4.6",
    "pillow-jxl-plugin>=1.3.2",
    "TextualLog @ git+https://github.com/HibernalGlow/TextualLog.git",
    "PresetUI @ git+https://github.com/HibernalGlow/PresetUI.git",
    "ttkbootstrap>=1.10.1",
    "portalocker>=2.7.0",
    "pynput>=1.7.6",
    "keyboard>=0.13.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
picsconvert = "picsconvert.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}  # 告诉 setuptools 包位于 src 目录
packages = ["picsconvert"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
ignore = ["E501"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.0.272",
    "black>=23.3.0",
]